+++ config/config.ini
@@ -1,0 +1,11 @@
+#UDAS Config Ini File.
+[Version]
+version=0.1
+
+[Management]
+auth_str=Passw0rd
+lang=en-US
+
+[Logging]
+path=.
+level=info
